requirements.txt
google-api-python-client==2.120.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
joblib==1.3.2
neo4j==5.19.0
py2neo==2021.2.4
streamlit==1.33.0
plotly==5.21.0
python-dotenv==1.0.1
matplotlib==3.8.4
seaborn==0.13.2
scipy==1.13.0
nltk==3.8.1
lime==0.2.0.1
pyvis==0.3.2
sentence-transformers==2.7.0
torch>=2.0.0
xgboost>=2.0.0
